+++ surnameGenerators/index.js
@@ -1,0 +1,7 @@
+const randomThirty = require('./random_thirty');
+const specific = require('./specific');
+
+module.exports = {
+    randomThirty,
+    specific
+};
