+++ surnameGenerators/specific.js
@@ -1,0 +1,10 @@
+const specificSurnames = [
+    '赵', '钱', '孙', '李', '周', '吴', '郑', '王'
+    // 可以在这里手动输入具体的姓氏
+];
+
+function getSpecificSurnames() {
+    return specificSurnames;
+}
+
+module.exports = getSpecificSurnames;
